test: use real AES cipher in an external test package; refine MAC comment
Address review feedback on the sequence-comment MAC fix:

- sops.go: replace the MAC-inclusion comment with the reviewer's wording
  explaining that non-Comment values can decrypt to Comment values, so
  isComment on the input cannot be used here.
- Move TestMACWithCommentInSequence into a new external test file
  (package sops_test), letting it import and use the real aes cipher. The
  previous stub cipher existed only because an internal (package sops) test
  cannot import aes without an import cycle (aes imports the root sops
  package). Drop the stub commentCipher.
- Use assert.False for the post-encrypt Comment check, and additionally
  assert the decrypted element is restored to a Comment with its content.
